requirements.txt
pyperclip
streamlit
requests
beautifulsoup4
bs4
st_copy_to_clipboard
pytesseract
PyPDF2
tesseract
Pillow
shutil